pico4ml/requirements.txt
pyserial>=3.5
requests>=2.0
pytest>=7.0
